Initial Commit; Readme & project requirements
This commit forms the initial commit and the basis of the repository, as
well outlines how the project will be executed from this point.

The commit should follow the pattern defined at:

  http://bit.ly/2kh67nV

== Design Notes

=== Version Control

While there is no specific requirement as part of this task to use
version control it is fairly fundamental part of the authors workflow.
Being able to traverse back and forward in time and use the primitives
exposed by git such as the commit, branching or other notions to enforce
application quality checks is super handy.

Commits should be "atomic" -- that is, a commit per a conceptual change
of a product. In this way when writing commit notes it is on a single
subject matter and when cherry-picking, reverting or otherwise
manipulating the git tree there is only a single logical change.

=== Documentation Heavy

This project should include documentation either as part of or before an
atomic commit.

In the authors experience the difficulty in software engineering is less
associated with a particular syntax, language choice or bug but rather
remembering how designs were arrived at over time and denoting under
which circumstances design decisions should be revisited.

The author likes designing software for the "longer term" (years).

=== Asciidoctor

The documentation is written in Asciidoctor (.ascii) format rather than
the more common markdown.

Asciidoctor exposes certain primitives that a particularly handy when
writing documentation such as callouts, crosslinks, diagrams and so
fourth.

Additionally, the toolchain for converting documents from asciidoctor
into other documents (html, pdf) is excellent.

More information can be found at the URL:

  https://l.littleman.co/34CxLOq

=== The Unseen

Not visible as part of the project history are the editor and other
aspects of the project that influence the authors work.

The current stack is:

- Linux on the desktop (Debain Sid)
- vscode as an editor
- terminal for the vast majority of other tasks