Initial layout of the project
This commit introduces the minimal available required to generate a
compilable entity on top of which further changes can be made and
documented.

== Stakeholder Impact
=== Developers

While there is only a single developer in this project, the requirements
hinted that there are other developers within the organisation that are
proficient in the language. This means the code is more accessible to
that team.

== Design Notes
=== Nonstandard go project layout

For go developers the layout of this project is a little unusual. The go
project layout is:

  https://l.littleman.co/2LPcEj8

That layout should be represented below the `src` folder where it is
associated with go source code.

However, the author generally works in a model where their teams own the
entire service down to the next logical API layer (usually Kubernetes).
This includes things like:

- CI/CD
- Production definitions
- Containers
- Documentation

It is possible to spread these across several different repositories.
However, in the authors experience the best "git" boundary reflects an
internal communications boundary and hopefully an API boundary. In this
case, those API boundaries are expected to be:

- The kubernetes API
- The upstream docker repository
- Service discovery via DNS or hard coded IP

More can be read at the following URL:

  https://l.littleman.co/2V6F6ns

=== Go modules

This commit uses the go modules system recently bundled with the core to
keep track of dependencies.

It further defines a non-standard name for this package, rather than the
traditional "github.com/${USERNAME}/${PASSWORD}". Because the
application is below the `src/` folder all libraries for this
application would already be ugly in the case that GitHub would be used
as the full package name.

Given this, choosing another package name that deliberately breaks
compatibility with the GitHub style is, in the authors mind, clearer.

=== sysexits.h

The only library in use currently is syexits.h. This library replicate
the "standard" exit codes initially in FreeBSD but also in sysexits.h in
Linux.

They should hopefully provide consumers a defined way to reason about an
applications termination state.