Create a dockerfile for application compilation and deployment
Golang has the "standard" application lifecycle management problems in
that it must be:

- Compiled to a deployable artifact
- Packaged in some sort of package manager
- Deployed onto production infrastructure

Docker provides excellent support to address all of these. This commit
introduces a docker image that allows the compilation of an image for
the binary.

== Stakeholder Impact
=== Developers

Given this design, developers do not need to worry about having a given
version of golang on their machine to compile the application. The
"correct" version of go will always be in a container.

=== Systems

Given this design the systems team do not need to worry about packaging
or understanding too much about the application. Rather, it can be
scanned for vulnerabilities and trust assigned to it via notary (either
by the developers or as part of a build pipeline) and the container
shipped to production with reasonable safety guarantees.

== Design Notes
=== Multi stage build

A multi stage docker image was used as the there are two competing goals
for this application package process:

1. The minimal application necessary for production, and
2. The dependencies required to package the application

Docker allows using one isolated environment for the compilation and
then the copying of objects from that environment to a "production like"
environment.

In this way, both needs are met.

=== No init process

In linux the process `pid 1` has some special responsibilities that are
not factored in to by the majority of processes. Specifically:

- process reaping
- signal propagation

In this case it is not anticipated this will be an issue with this
process. There is nothing in the production environment for golang to
spawn that might need reaping and go has sane defaults for how it
handles unix signals such as SIGTERM, SIGINT. See:

  See https://golang.org/src/os/signal/doc.go

=== gcr.io

The tag specified in the image documentation is "gcr.io". This is the
Google Cloud container registry, and the canonical registry the author
uses for pet projects.

It provides cheap storage, vulnerability scanning and a solid access
model with Google Cloud IAM.