Add a minimal CI/CD configuration
Generally speaking the author will add the CI/CD configuration as early
in the project as possible. This is for several reasons:

1. The infrastructure affects application engineer decisions, and
   allowing engineers to verify their changes against production like
   infrastructure early saves later integration pain
2. Standards enforced in the CI process are very hard to achieve buy in
   for in retrospect

== Design Notes

=== Drone

While covered more in the README, drone.io is the authors canonical
choice for CI/CD. It is docker driven, declarative and opinionated.

=== `img`

Within the pipeline the tool "img" is used to build the container,
rather than the canonical docker.

The problem is in the execution environment of drone -- that is, on a
Kubernetes cluster. The choices for building images in a Kubernetes
cluster are either:

1. Mount in the docker socket
2. Find a way to build images in an unprivileged way

The former approach immediately opens up the cluster for privilege
escalation via the mounted in docker socket -- effectively, it is
immediate RCE to root.

The `img` tooling is by Jessie Frazelle, a former docker / kubernetes
maintainer, has no daemon and allows building images quickly and easily.
It requires minimal configuration to allow it to work on Debian kernels
as Debian modifies the default user namespace behaviour.

=== Commit hooks

Because this project is not in CI and there will be no "pull request"
workflow the things that are ordinarily enforced via pull request are
instead enforced via git hook.

The workflow is largely the same.

=== golangci-lint

This commit introduces the linting tool "golangci-lint". This is an
aggregator, and runs a whole series of linting tools against go code.

While go code is usually fairly strict, the golangci-lint should enforce
a very high coding standard indeed -- particularly for intermediate go
programmers such as the author.