Increase timelines for lints
In order to do the static analysis on the application the tool needs to
vendor all application dependencies.

There are a few of these, so it takes some period of time -- more than
the default 1m that the tool expects.

This commit bumps the wait to 5 minutes. In parallel investigatory work
is being done to introduce a caching layer to the builds.