Sketch out the application skeleton with spf13/cobra
Currently there is not so much in the way of application structure. It
would be possible to use Go's native flag parsing library and structure
the application without former precedent, however the author has
implemented several small Go CLI programs previously with the spf13/cobra
and spf13/viper toolkit.

Those programs take care of some initial issues, including:

- A CLI interface to the application that is logical and feels \
  familiar (the same interface is in kubectl, hugo etc)
- A configuration paradigm and parsing engine

Both of which are extremely common issues for an application to need to
resolve.

Accordingly, this commit introduces the minimum required to bootstrap
that toolkit.