Introduce caching layer into the CI process
Currently there several build steps associated with invoking, building,
testing or otherwise manipulating the go binary.

These are defined separately such that they can form a directed acyclic
graph (DAG) in as many different formations as the graph as possible.
However, currently there is no way that go dependencies are transferred
across steps, meaning the vendored dependencies are revendored for every
step.

This commit addresses that by introducing a caching layer that simply
writes the cache to a temporary directory on the host.

== Design Notes
=== Security

The implication to caching content on the host is that users who control
the build are then able to write content on the host. This might include
suid binaries that allow privilege escalation.

This is a bad idea.

It is less a bad idea to write it to `/tmp` than to other directories
as `/tmp` is hopefully mounted `noexec,nosuid` in `/etc/fstab`. However,
in all probability that is not the case.

Instead, some sort of storage like S3 should be used for this caching
layer. In this project that is deemed unnecessary as there will likely
only ever be a single contributor, running builds on their local
machine.