Add version command
With any software there is a gradual evolution over time improving the
software, reducing bugs or otherwise making changes to how the software
works.

The way to track that primitive is the "version" of the software that is
used.

This commit adds a command to describe the version in both an
abbreviated and extended way.

== Design Notes

=== Generated version information

This commit generates the version information from the git history,
rather than hard coding it. This seems to be the simplest way to
consistently get correct information in the application release process.

=== Defaults

The defaults of the release process should clearly indicate when this is
an "out of band" release, or a build used in development only.