Implement a stub connection handler
The broad goal of this work is to allow connections from entities over
TCP for DNS and have that forwarded to an upstream DNS server.

This commit goes some way to implementing this by reading the DNS packet
from the incoming connection. It does not implement any logic associated
beyond determining the appropriate length of the packet and dumping a
hex representation to standard out.

== Design Notes

=== TCP specific handler

Over TCP the DNS protocol payload is slightly different than that over
UDP. Specifically, the DNS message is prefixed with a 2 byte length
property that indicates how long the DNS message will be. UDP, by
comparison, is simply cut off at 512 bytes.

The current handler implementation is becoming very TCP specific. Future
work should refactor it out such that reading in the packet payload is a
distinct step from assembling and understanding that packet.

== Design Notes

=== General Design

The design is largely inspired by:

  https://golang.org/pkg/net/#example_Listener

== References

The following upstream references were particularly useful while writing
this work:

- https://gist.github.com/hongster/04660a20f2498fb7b680
- https://golang.org/pkg/net/#Conn
- https://medium.com/go-walkthrough/go-walkthrough-encoding-binary-96dc5d4abb5d