Add a logging library
Within the application there are various print statements to determine
the internal state of the program, useful when trying to understand
the applications internal state.

However, these statements are unstructured. When deployed to a
production environment such statements are difficult to correlate
against other events over time, as well difficult to parse for deeper
insights into the application.

This commit addresses that by adding the "logarus" library to the
application; a common library to handle just this problem.

== Design Notes

=== JSON logs

By default, this application logs its output in JSON instead of a
structured text format.

The thinking is that in the majority of cases the logs will be consumed
by error detection machinery, rather than humans. Further, humans can
easily drop the logs back to a single text format with software such as
`jq`:

  ./pdns | jq '.msg'

=== WithFields

Rather than interpolating a string with variables the logging library
encourages embedding relevant data in a struct that is attached to the
JSOB object, but not as part of the message.

This allows a simpler indexing and query path for the logs.