Functional prototype of the proxy
This commit introduces a functional prototype of the proxy that just
forwards the responses to the DNS over TLS proxy

== Design Notes

=== Initial Spec

Initially the reading of the project requirements lead the author to
understand that the DNS queries should be unpacked an represented in an
in-software primitive (struct), before being reserialized and forwarded
on to the intermediary proxy.

That still seems the best implementation path as it allows introspection
of the requests and adding of a caching layer; both early requirements
of most DNS systems.

However, reimplementing the DNS spec proved expensive in terms of time.
Given this, the minimal implementation has been created that adheres to
the requirements.

== Known Bugs

=== Additional Padding

The CloudFlare DNS over TLS resolver is returning an incorrect length
prefix for responses, giving additional return for responses that do
not support it.

See https://l.littleman.co/2LTRyjK