Push the built image to GitHub
Currently while the image is built in the pipeline it is not pushed into
any artifact repository. The author has used the following:

- quay.io
- gcr.io
- docker hub

However, they have not yet used the recently released (to beta) GitHub
package repository. This seems like a good opportunity to test that.

== Design Notes
=== Secrets

This requires authenticating against GitHub. This is done via using a
personal access token as the GitHub account has two factor on it and
even should it not, giving shared credentials to third party services
has nasty failure domains regardless.

Initially the secret design for this commit included `git-crypt`.
However, in principle it is better to defer credential management to a
third party identity provider that has strong authentication guarantees
itself (i.e. Vault), and the closest implementation to that is to supply
the variables to Drone which then grants access to them in specific
stages of the build.

The variables are documented in `.env` for local execution.