Restructure the server API
While the application "works" currently it is not laid out in such a
way that it is easy to understand or configure. This commit attempts
to improve that, specifically for the server object.

The next feature that will be introduced is the ability to customize the
upstream resolver for the application. Further, it is extremely likely
that the server will be further customizable in future; things such as:

- listen port
- listen address
- upstream resolver port
- protocol
- udp support

And so on. There is a limited amount of expected customizability in
these features which needs to be supplied to the server entity on its
instantiation.

This commit attempts to address that by refactoring the server to be a
struct type and require a constructor. Further, this constructor now
accepts (but does not use) a configuration object that will later be
used to define the expected behaviour of the application.

Within this commit, the functionality of the application is not changed.
Rather, the APIs that are used to create and manage objects are.

Future work will continue to refine this API, as well as to introduce
configurable components of the application.

== References

1. https://dave.cheney.net/2014/10/17/functional-options-for-friendly-apis