Replace GitHub with Quay
Unfortunately it appears that there is a compatibility issue with the
`img` build tool and the github implementation of the docker repository.
More specifically, the repository returns:

  HTTP 406 Not Acceptable

Whenever an image upload is attempted. Rather than spend the time to
reverse engineer that big, this commit just sidesteps around it by using
another, more stable registry.

== Security
=== Open Image

This does mean that the pdns image is now available to all by default.
However, given that it is compiled and that it is not a particularly
secret project, this should have no negative consequences.