Switch to gcr.io
This commit introduces yet another attempt to find an upstream
repository that works with the `img` container builder.

The issue is better explained for `quay.io`, in that it does not support
the `v2` schema for defining images:

  https://github.com/moby/buildkit/issues/409

It seems likely that Docker suffers a similar issue. This toolchain has
previously been used with gcr and is expected to work correctly with it
now.

== Design Notes
=== Still weirdly broken

In order to have this work correctly the image first had to be pushed
with docker, then was supported by the `img` tool. Strange.

=== Surprising

It is surprising if it is the case that few registries support v2, and
more surprising still that Docker is using the deprecated v1 at release.