Shift starting the server to a separate command
Currently when the application is run by default it will start the
server and listen for connections. However, there is a lot of logic in
the initial "rootCommand" method, and no easy way to guide people
through the execution of the program.

This commit shifts booting the server to a separate subcommand in which
more logic associated with creating the server can go without making the
program "messy" or more difficult to read.

The root command exits nonzero by default to indicate to users that this
command does not have a useful action.

BREAKING CHANGE: This changes the CLI API and scripts that simply invoke
                 the binary will no longer work.