Allow selecting the upstream provider
There are several possible upstream providers that provide DNS over TLS,
including:

- 8.8.8.8
- 1.1.1.1
- 8.8.4.4

It stands to reason the consumers of this project should be able to
determine which provider they wish to use.

== Design Notes
=== Configuration

At the time of writing the author is still not super sure what a good
object configuration API looks like. The current design provides a
default configuration struct via constructor and allows overriding that
with a method on that type; it should work in the vast majority of cases
and limit the amount of footgunnery that happens, but the API does not
feel super graceful to work with.

More thought here is required.