Rework server configuration API, prototype udp listener
The next feature to be implemented is to provide a parallel UDP listener
on which the DNS server is exposed.

This is interesting in a couple of respects:

- The application needs to be bound on both TCP and UDP ports. That is,
  it needs to expose two listeners
- UDP is not length prefixed, and thus needs a logical rework
- UDP is not yet implemented

Further, there are other protocols that could be implemented but are not
yet.

This commit implements a skeleton for implementing each of those
protocol. It further reworks the configuration such that it can validate
correctness for a given set of configuration, but is still extensible.

== Known Limitations
=== Host role & Protocol

Currently the host role allows both "dot" and "tcp" for both ends of the
connection. This is not correct; the proxy only implements TCP and does
not allow pushing forward with TCP connections to upstream.

Future work should reconcile the ends of the proxy such that they use
the same protocol and serialization process and thus have parity in
their protocol support.

=== Multiple error conditions

Because there is now validation on the configuration creation there is a
swathe of error conditions that are handled individually.

It would be possible to put this into a function, but feels a little
unclear in terms of what that abstraction is supposed to represent.

More thinking required.