Implement UDP handler
In previous work a TCP listener was implemented in the DNS proxy. This
listener is fairly simple to implement as the payload can be read
directly from the incoming TCP handler and written back to TCP enveloped
in TLS, and the response reverse proxied back.

However, the UDP message is slightly different. TCP requires a length
prefix for all message, however UDP does not as the packet can be
directly read off the stack and processed.

Given this, the network handling logic needs to be implemented slightly
differently for each protocol before the payload is massaged into a
given structure suitable for sending on to the DLS over TLS
implementation before being translated back for the response.

== Design Notes
=== Mixed responsibilities

The file  "server.go" now handles both TLS and UDP connections, as well
as the shared functionality therein. It is a good candidate to be
restructured such that the common functionality is split out, and the
unique per-connection functionality split out.