Add privilege reduction to containers
Within Linux a process is started within a given namespace as "root",
and unless it deliberately sheds its privileges it will remain as root.
Further, without user namespace remapping (not stable in Kubernetes) the
user within the container is the same as the user on the host -- root is
still root.

Given this, the best practice for containers is to run processes within
that container as a user who has a limited privilege.

The author has not yet seen this implemented in the "scratch"
containers, however it stands to reason it is possible to run processes
as a lower privilege user so long as the runtime understands who that
user is.

This commit attempts to implement this by:

- Granting the process setcap privileges so it can still bind to port
  53
- Running the process as the "daemon" user (1), a less privileged user.