Allow listening on all ports, usable docker image
Currently the pdns application is hard-coded to only listen on
localhost. The thinking behind that decision was that it is not
reasonable to answer queries not intended for a given destination, and
the primary use case anticipated for this resolver would be:

1. The sidecar deployment (in which case 127.0.0.1 is fine)
2. On a host (the same)

In both cases the network namespace is shared and all content available
over localhost.

However, currently when the docker image is used without the sidecar
pattern it is a little useless as it does not answer queries outside
its namespace, and given the design of docker everything will be outside
its namespace.

This commit addresses that by allowing configuration of the listen
address as well as setting the dockerfile to listen on all adapters by
default.